refactor(design-system-impl): promote closed-tab stack to reactive .svelte.ts (1.9b0)
Back the closed-tab LIFO with `$state` and expose a reactive `top` getter so the
a7 launcher's "Reopen last closed" affordance can enable/disable live while the
menu is open. Imperative push/pop/peek/size/clear surface is unchanged — App's
Ctrl+Alt+T flow and the existing unit tests keep working verbatim.

Isolated commit so a regression bisects away from the launcher UI rebuild.
